edit : 유효하지 않은 페이지 번호 입력 시 범위 내의 번호로 설정

+++ Back-end/api-module/src/main/java/com/ico/api/dto/treasuryHistory/TreasuryHistoryTeacherResDto.java
@@ -15,12 +15,12 @@ import java.util.List;
 @NoArgsConstructor
 public class TreasuryHistoryTeacherResDto {
 
-    private long size;
+    private int size;
 
     private List<TreasuryHistoryTeacherColDto> page;
 
     @Builder
-    public TreasuryHistoryTeacherResDto(long size, List<TreasuryHistoryTeacherColDto> page) {
+    public TreasuryHistoryTeacherResDto(int size, List<TreasuryHistoryTeacherColDto> page) {
         this.size = size;
         this.page = page;
     }
